+++ ports/signalrclient/portfile.cmake
@@ -11,17 +11,17 @@ vcpkg_extract_source_archive(${ARCHIVE})
 vcpkg_apply_patches(
     SOURCE_PATH ${SOURCE_PATH}
     PATCHES
-		${CMAKE_CURRENT_LIST_DIR}/0001_cmake.patch
+        ${CMAKE_CURRENT_LIST_DIR}/0001_cmake.patch
 )
 
 vcpkg_configure_cmake(
     SOURCE_PATH ${SOURCE_PATH}
-	OPTIONS_DEBUG
-		-DCPPREST_SO=${CURRENT_INSTALLED_DIR}/debug/lib/cpprest_2_9.lib
-	OPTIONS_RELEASE
-		-DCPPREST_SO=${CURRENT_INSTALLED_DIR}/lib/cpprest_2_9.lib
+    OPTIONS_DEBUG
+        -DCPPREST_SO=${CURRENT_INSTALLED_DIR}/debug/lib/cpprest_2_10.lib
+    OPTIONS_RELEASE
+        -DCPPREST_SO=${CURRENT_INSTALLED_DIR}/lib/cpprest_2_10.lib
     OPTIONS
-		-DCPPREST_INCLUDE_DIR=${CURRENT_INSTALLED_DIR}/include
+        -DCPPREST_INCLUDE_DIR=${CURRENT_INSTALLED_DIR}/include
         -DDISABLE_TESTS=ON
 )